requirements.txt
google-genai
groq
openai
huggingface-hub
